pyproject.toml
[project]
name = "tinywars"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.7",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
    "ty>=0.0.20",
]

[tool.ruff]
line-length = 92
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E303"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
force-sort-within-sections = true